Cleanup file structure
